feat: added router as dynamic param for custom UniswapV2Library with dex own fees

+++ contracts/libraries/UniswapV2Library.sol
@@ -16,7 +16,11 @@ library UniswapV2Library {
     }
 
     // calculates the CREATE2 address for a pair without making any external calls
-    function pairFor(address factory, address tokenA, address tokenB) internal view returns (address pair) {
+    function pairFor(
+        address factory,
+        address tokenA,
+        address tokenB
+    ) internal view returns (address pair) {
         return IUniswapV2Factory(factory).getPair(tokenA, tokenB);
     }
 
@@ -32,7 +36,11 @@ library UniswapV2Library {
     }
 
     // given some amount of an asset and pair reserves, returns an equivalent amount of the other asset
-    function quote(uint amountA, uint reserveA, uint reserveB) internal pure returns (uint amountB) {
+    function quote(
+        uint256 amountA,
+        uint256 reserveA,
+        uint256 reserveB
+    ) internal pure returns (uint256 amountB) {
         require(amountA > 0, 'INSUFFICIENT_AMOUNT');
         require(reserveA > 0 && reserveB > 0, 'INSUFFICIENT_LIQUIDITY');
         amountB = amountA.mul(reserveB) / reserveA;
@@ -80,7 +88,11 @@ library UniswapV2Library {
         }
     }
 
-    function createPair(address factory, address tokenA, address tokenB) internal returns (address pair) {
+    function createPair(
+        address factory,
+        address tokenA,
+        address tokenB
+    ) internal returns (address pair) {
         return IUniswapV2Factory(factory).createPair(tokenA, tokenB);
     }
 }
